Namespace and class names have been revised.
