+++ src/Carbunql.TypeSafe/Building/BuilderEngine.cs
@@ -1,25 +1,24 @@
 ﻿using Carbunql.Building;
 using Carbunql.Extensions;
+using Carbunql.TypeSafe.Dialect;
 using System.Data;
 
-namespace Carbunql.TypeSafe;
+namespace Carbunql.TypeSafe.Building;
 
-internal class BuilderEngine(ISqlDialect sqlDialect, IParameterManager manager)
+internal class BuilderEngine(ISqlDialect sqlDialect, IConstantRegistry registory)
 {
     internal readonly ISqlDialect SqlDialect = sqlDialect;
-    public readonly IParameterManager Manager = manager;
-
-    //public IEnumerable<KeyValuePair<string, object>> Parameters => Manager.Parameters;
+    public readonly IConstantRegistry Registory = registory;
 
     public string AddParameter(string name, object? value)
     {
         if (string.IsNullOrEmpty(name)) throw new ArgumentNullException(nameof(name));
 
         var key = SqlDialect.ToParaemterName(name);
-        var q = Manager.Parameters.Where(x => key.IsEqualNoCase(x.Key));
+        var q = Registory.Parameters.Where(x => key.IsEqualNoCase(x.Key));
         if (!q.Any())
         {
-            return Manager.AddParameter(key, value);
+            return Registory.Add(key, value);
         }
         else if (q.First().Value.Equals(value))
         {
@@ -37,22 +36,17 @@ internal class BuilderEngine(ISqlDialect sqlDialect, IParameterManager manager)
     /// <exception cref="InvalidProgramException"></exception>
     public string AddParameter(object value)
     {
-        var q = Manager.Parameters.Where(x => x.Value != null && x.Value.Equals(value));
+        var q = Registory.Parameters.Where(x => x.Value != null && x.Value.Equals(value));
         if (q.Any())
         {
             return q.First().Key;
         }
         else
         {
-            var name = $"const_{Manager.Parameters.Count()}";
+            var name = $"const_{Registory.Parameters.Count()}";
             return AddParameter(name, value);
         }
 
         throw new InvalidProgramException();
     }
-
-    //public int Count()
-    //{
-    //    return Manager.Count();
-    //}
 }
