--- src/Carbunql.TypeSafe/CTEDefinition.cs
@@ -1,10 +1,0 @@
-﻿namespace Carbunql.TypeSafe;
-
-public struct CTEDefinition
-{
-    public string Name { get; set; }
-
-    public Type RowType { get; set; }
-
-    public string Query { get; set; }
-}
