+++ src/Carbunql.TypeSafe/Extensions/BinaryExpressionExtension.cs
@@ -1,5 +1,6 @@
 ﻿using Carbunql.Analysis.Parser;
 using Carbunql.Extensions;
+using Carbunql.TypeSafe.Building;
 using System.Linq.Expressions;
 
 namespace Carbunql.TypeSafe.Extensions;
