+++ src/Carbunql.TypeSafe/FluentSelectQuery.cs
@@ -5,6 +5,8 @@ using Carbunql.Clauses;
 using Carbunql.Definitions;
 using Carbunql.Extensions;
 using Carbunql.Tables;
+using Carbunql.TypeSafe.Building;
+using Carbunql.TypeSafe.Dialect;
 using Carbunql.TypeSafe.Extensions;
 using Carbunql.Values;
 using System.Data;
@@ -31,7 +33,7 @@ public class FluentSelectQuery : SelectQuery
         var analyzed = ExpressionReader.Analyze(expression);
 #endif
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         if (expression.Body is MemberExpression mem)
@@ -134,7 +136,7 @@ public class FluentSelectQuery : SelectQuery
         var analyzed = ExpressionReader.Analyze(expression);
 #endif
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         if (expression.Body is NewExpression ne)
@@ -168,7 +170,7 @@ public class FluentSelectQuery : SelectQuery
         var table = compiledExpression();
 
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var condition = conditionExpression.Body.ToValue(engine);
@@ -191,7 +193,7 @@ public class FluentSelectQuery : SelectQuery
         var table = compiledExpression();
 
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var condition = conditionExpression.Body.ToValue(engine);
@@ -214,7 +216,7 @@ public class FluentSelectQuery : SelectQuery
         var table = compiledExpression();
 
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var condition = conditionExpression.Body.ToValue(engine);
@@ -245,7 +247,7 @@ public class FluentSelectQuery : SelectQuery
 #endif
 
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var value = expression.Body.ToValue(engine);
@@ -261,7 +263,7 @@ public class FluentSelectQuery : SelectQuery
         var analyzed = ExpressionReader.Analyze(expression);
 #endif
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var value = expression.Body.ToValue(engine);
@@ -283,7 +285,7 @@ public class FluentSelectQuery : SelectQuery
         var analyzed = ExpressionReader.Analyze(expression);
 #endif
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var value = expression.Body.ToValue(engine);
@@ -307,7 +309,7 @@ public class FluentSelectQuery : SelectQuery
         var analyzed = ExpressionReader.Analyze(expression);
 #endif
         var parameters = QueryParameterMerger.Merge(GetParameters());
-        var prmManager = new ParameterManager(parameters, AddParameter);
+        var prmManager = new ParameterRegistory(parameters, AddParameter);
         var engine = new BuilderEngine(SqlDialect, prmManager);
 
         var value = expression.Body.ToValue(engine);
@@ -327,7 +329,7 @@ public class FluentSelectQuery : SelectQuery
         var analyzed = ExpressionReader.Analyze(expression);
 #endif
 
-        var prmManager = new ParameterManager(GetParameters(), AddParameter);
+        var prmManager = new ParameterRegistory(GetParameters(), AddParameter);
 
         if (expression.Body is not MemberExpression me)
         {
