+++ src/Carbunql.TypeSafe/FluentSelectQueryBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq.Expressions;
+using Carbunql.TypeSafe.Dialect;
 
 namespace Carbunql.TypeSafe;
