+++ test/Carbunql.TypeSafe.Test/ExistsTest.cs
@@ -1,5 +1,4 @@
-﻿using Carbunql.Building;
-using Xunit.Abstractions;
+﻿using Xunit.Abstractions;
 
 namespace Carbunql.TypeSafe.Test;
