+++ test/Carbunql.TypeSafe.Test/SqlDialectTest.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Carbunql.TypeSafe.Dialect;
+using Microsoft.Extensions.DependencyInjection;
 using Xunit.Abstractions;
 
 namespace Carbunql.TypeSafe.Test;
